+++ test/unit/check_log/fetch_log.py
@@ -43,6 +43,7 @@ class ArgParser(object):
 
     Methods:
         __init__
+        get_val
 
     """
 
@@ -57,6 +58,18 @@ class ArgParser(object):
         """
 
         self.args_array = dict()
+
+    def get_val(self, skey, def_val=None):
+
+        """Method:  get_val
+
+        Description:  Method stub holder for gen_class.ArgParser.get_val.
+
+        Arguments:
+
+        """
+
+        return self.args_array.get(skey, def_val)
 
 
 class UnitTest(unittest.TestCase):
@@ -153,6 +166,7 @@ class UnitTest(unittest.TestCase):
         """
 
         self.argspar.args_array = {"-f": [self.file_0]}
+
         mock_open.return_value = open(self.argspar.args_array["-f"][0], "r")
 
         check_log.fetch_log(self.log, self.argspar)
@@ -171,6 +185,7 @@ class UnitTest(unittest.TestCase):
         """
 
         self.argspar.args_array = {"-f": [self.file_1]}
+
         mock_open.return_value = open(self.argspar.args_array["-f"][0], "r")
 
         check_log.fetch_log(self.log, self.argspar)