+++ test/unit/check_log/main.py
@@ -49,6 +49,7 @@ class ArgParser(object):
         arg_exist
         arg_file_chk
         arg_valid_val
+        get_val
 
     """
 
@@ -164,6 +165,18 @@ class ArgParser(object):
         self.opt_valid_val = opt_valid_val
 
         return self.arg_valid_val2
+
+    def get_val(self, skey, def_val):
+
+        """Method:  get_val
+
+        Description:  Method stub holder for gen_class.ArgParser.get_val.
+
+        Arguments:
+
+        """
+
+        return self.args_array.get(skey, def_val)
 
 
 class ProgramLock(object):