+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [4.0.1] - 2022-05-06
+### Changed
+- run_program, load_attributes, fetch_log, update_marker, full_chk: Replaced direct access of args_array with get_val call.
+- main, log_2_output: Replaced get call from args_array with method call to get_val.
+
+
 ## [4.0.0] - 2022-02-07
 Breaking Changes